Sockets are closed on Processor.shutdown()
Before a socket is closed, the setsockopt(zmq.LINGER, 0)
option is set to force the socket to close, regardless of
pending messages. This allows Processors to be started up
multiple times in the same parent python program.

If Processor.shutdown() is not closed, PyZMQ will complain that
sockets are already in use.